Anemonode v1.0-rc2

+++ src/device/anemobox/anemonode/install.sh
@@ -1,32 +1,33 @@
 #!/bin/bash
 set -e
 
-HOST=192.168.1.109
+# TODO: get the IP from the command line
+HOST=192.168.1.106
 
-rm -fR node_modules/mail node_modules/endpoint || true
-export NODE_ENV=production
-npm install
+# avoid rebuilding, because rebuilding will make npm updates the dependencies
+# in node_modules, which in turn will lead to a large patch
+if false; then
 
-node-gyp configure
-node-gyp build
+  rm -fR node_modules/mail node_modules/endpoint || true
+  export NODE_ENV=production
+  npm install
 
-npm prune --production
-npm dedupe
+  node-gyp configure
+  node-gyp build
 
-../canutils/compile-can-utils.sh
-../canutils/compile-iproute2.sh
+  npm prune --production
+  npm dedupe
 
-EXCLUDE='--exclude=".*.sw[po]" --exclude="src" --exclude="*\.o" --exclude="binding.gyp" --exclude=README.md --exclude=install.sh --exclude=build/Release/obj'
+  ../canutils/compile-can-utils.sh
+  ../canutils/compile-iproute2.sh
+fi
+
+EXCLUDE='--exclude=*.log --exclude=.*.sw[po] --exclude=src --exclude=*\.o --exclude=*.o.d --exclude=binding.gyp --exclude=README.md --exclude=install.sh --exclude=build/Release/obj'
 
 git rev-parse HEAD > commit
 
 ssh root@${HOST} rm -fR "/anemonode/*"
 rsync -ar ${EXCLUDE} . root@${HOST}:/anemonode
-
-#[ -e /anemonode ] || git clone git+ssh://anemobox@vtiger.anemomind.com/home/anemobox/anemobox.git /anemonode
-#rm -fR /anemonode/*
-#rsync -ar --exclude="src/*" --exclude="*.o" --exclude=id_rsa --exclude=install.sh . /anemonode
-
 
 echo "Installed. After testing, please validate the release files with: "
 echo "  cd /anemonode ; git add . ; git commit ; git push"
